% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/dmixst.R
\name{dmixst}
\alias{dmixst}
\title{The value of the mixed densities of multivarite skew t distributions.}
\usage{
dmixst(x, para, log = FALSE)
}
\arguments{
\item{x}{a vector.}

\item{para}{an object of the parameters of the mixture of multivariate skew t distributions.
pro a vector for the mixture ratios, mu the values, sigma, delta, nu}

\item{log}{logical; if TRUE, probabilities p are given as log(p).}
}
\value{
numeric vector with the values of the mixture of the density functions.
}
\description{
The value of the mixed densities of multivarite skew t distributions.
}
\examples{
x=c(1,2)
pro = 1; mu=1; Sigma=3; delta=2; nu=10;
para = list(pro=pro,mu=mu,Sigma=Sigma,delta=delta,nu=nu)
dmixst(x,para)
}
\keyword{density}
\keyword{function}

