% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{fitmixout-class}
\alias{fitmixout-class}
\title{An S4 class the stores the S4 class mixpara, the input data, logLik, S2, S3, iter and call}
\description{
An S4 class the stores the S4 class mixpara, the input data, logLik, S2, S3, iter and call
}
\keyword{classes}

