% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/logLik-method.R
\docType{methods}
\name{logLik-method}
\alias{logLik,fitmixout-method}
\alias{logLik-method}
\title{extract plot}
\usage{
\S4method{logLik}{fitmixout}(object)
}
\description{
extract plot
}

