% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{mixpara-class}
\alias{mixpara-class}
\title{An S4 class the stores the parameters of a mixed skew t distribution}
\description{
An S4 class the stores the parameters of a mixed skew t distribution
}
\keyword{classes}

