% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/mmixstEstC.R
\name{mmixstEstC}
\alias{mmixstEstC}
\title{Internal fitting function}
\usage{
mmixstEstC(y, g, itermax, error, pro, mu, Sigma, delta, nu, verbose = F)
}
\arguments{
\item{y}{a multidimensional input vector.}

\item{g}{the number of groups.}

\item{itermax}{the maximum of iterations. (default 1000)}

\item{error}{the desired relative error. (default 1e-5)}

\item{pro}{the number of groups.}

\item{mu}{the desired relative error. (default 1e-5)}

\item{Sigma}{the maximum of iterations. (default 1000)}

\item{delta}{of finding initial values. (default "kmeans")}

\item{nu}{of finding initial values. (default "kmeans")}

\item{verbose}{of finding initial values. (default "kmeans")}
}
\value{
a object of the class fitmixst.
}
\description{
Internal fitting function
}
\keyword{fit}
\keyword{mixed}
\keyword{skew}
\keyword{t.}

