% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plot-method.R
\docType{methods}
\name{plot-method}
\alias{lines,fitmixout-method}
\alias{plot,fitmixout,ANY-method}
\alias{plot,fitmixout-method}
\alias{plot,mixpara-method}
\alias{plot-method}
\alias{points,fitmixout-method}
\title{extract plot}
\usage{
\S4method{plot}{mixpara,ANY}(x, y, xlab = "x-axis", ylab = "density",
  from = -1, to = 1, ...)

\S4method{plot}{fitmixout,ANY}(x, y, xlab = "x-axis", ylab = "density",
  main = x@call, type = "l", from = min(x@y), to = max(x@y), ...)

\S4method{lines}{fitmixout}(x, xlab = "x-axis", ylab = "density", ...)

\S4method{points}{fitmixout}(x, ...)
}
\description{
extract plot
}

