% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/predict-method.R
\docType{methods}
\name{predict-method}
\alias{predict,fitmixout-method}
\alias{predict-method}
\title{extract probability}
\usage{
\S4method{predict}{fitmixout}(object)
}
\description{
extract probability
}

