% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/regSkewt.R
\name{regSkewt}
\alias{regSkewt}
\title{Fitted mixed model}
\usage{
fitmixst(y, g, rel.error = 1e-3, itermax=1000, method = "kmeans", para = NULL, verbose = F, mcmc = F, useR = F, ...)
}
\arguments{
\item{y}{a multidimensional input vector.}

\item{rel.error}{the desired relative error. (default 1e-3)}

\item{itermax}{the maximum of iterations. (default 1000)}

\item{method}{of finding initial values. (default "kmeans")}

\item{para}{the initial input parameters. Note: if you choose input parameters the kmeans method is disabled.
para a list of the input parameters. para = list (pro, mu, Sigma, delta, nu).}

\item{verbose}{debug mode.}

\item{mcmc}{calculates moments moments with mcmc in R}

\item{useR}{use algorithm programmed in R (default is C++ Algorithm).}

\item{R2}{use alternative R code.}

\item{...}{other paramters.}

\item{g}{the number of groups.}
}
\value{
an S4 object of the class fitmixout.
}
\description{
Fitted mixed model
}
\examples{
pro = 1; mu=1; Sigma=3; delta=2; nu=3;
para = list(pro=pro,mu=mu,Sigma=Sigma,delta=delta,nu=nu)
y <- rmixst(100,para)
out <- fitmixst(y,g=1,method="kmeans")
}
\keyword{distribution.}
\keyword{fit}
\keyword{skew}
\keyword{t}

