% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/rmvn.R
\name{rmvn}
\alias{rmvn}
\title{Generates random observation for the normal distribution}
\usage{
rmvn(n, p, mean = rep(0, p), cov = diag(p))
}
\arguments{
\item{n}{number of random observations.}

\item{p}{dimension of the normal distribution.}

\item{mean}{a mean vector with length p.}

\item{cov}{a covariance matrix of dimension pxp.}
}
\value{
gernerates random observations for the normal distribution.
}
\description{
Generates random observation for the normal distribution
}
\examples{
mu=1; Sigma=1;
y <- rmvn(n=100,p=1,mean=mu,cov=Sigma)
}
\keyword{distribution.}
\keyword{normal}
\keyword{observations}
\keyword{random}

