% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/rmvnorm.R
\name{rmvnorm}
\alias{rmvnorm}
\title{Generates random observations for the normal distribution}
\usage{
rmvnorm(n, mean = rep(0, nrow(sigma)), sigma = diag(length(mean)),
  method = c("eigen", "svd", "chol"))
}
\arguments{
\item{n}{number of random observations.}

\item{mean}{a mean vector with length p.}

\item{sigma}{a covariance matrix of dimension pxp.}

\item{method}{used for inverse (eigen, svd, chol).}
}
\value{
gernerates random observations for the normal distribution.
}
\description{
Generates random observations for the normal distribution
}
\keyword{distribution.}
\keyword{normal}
\keyword{observations}
\keyword{random}

