% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/show-method.R
\docType{methods}
\name{show-method}
\alias{show,fitmixout-method}
\alias{show,mixpara-method}
\alias{show-method}
\title{extract show}
\usage{
\S4method{show}{mixpara}(object)

\S4method{show}{fitmixout}(object)
}
\description{
extract show
}

