% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/stderror.R
\name{stderror}
\alias{stderror}
\title{The value of the mixed densities of multivarite skew t distributions.}
\usage{
stderror(x, model, log = FALSE)
}
\arguments{
\item{x}{grid for which the standard errors are calculated}

\item{model}{an object of the parameters of the mixture of multivariate skew t distributions.
pro a vector for the mixture ratios, mu the values, sigma, delta, nu}

\item{log}{logical; if TRUE, probabilities p are given as log(p).}
}
\value{
numeric vector with the values of the mixture of the density functions.
}
\description{
The value of the mixed densities of multivarite skew t distributions.
}
\examples{
pro = 1; mu=1; Sigma=3; delta=2; nu=10;
para = list(pro=pro,mu=mu,Sigma=Sigma,delta=delta,nu=nu)
y=rmixst(100,para)
out <- fitmixst(y,g=1,method="kmeans")
std = stderror(y, out)
}
\keyword{error}
\keyword{function}
\keyword{standard}

