% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/summary-method.R
\docType{methods}
\name{summary-method}
\alias{summary,fitmixout-method}
\alias{summary-method}
\title{extract summary}
\usage{
\S4method{summary}{fitmixout}(object)
}
\description{
extract summary
}

