% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/truncatedt.R
\name{truncatedt}
\alias{truncatedt}
\title{Internal function for moments of truncated t}
\usage{
truncatedt(a, mu, sigma, nu)
}
\arguments{
\item{a}{truncation at}

\item{mu}{mean of t distribution}

\item{sigma}{scale of t distribution}

\item{nu}{degrees of freedom}
}
\value{
first and second moment of the truncated t distribution
}
\description{
Internal function for moments of truncated t
}
\keyword{distribution}
\keyword{t}
\keyword{truncated}

